+++ packages/package-manager/src/NodePackageManagers.ts
@@ -1,7 +1,6 @@
 import JsonFile from '@expo/json-file';
 import spawnAsync, { SpawnOptions } from '@expo/spawn-async';
 import ansiRegex from 'ansi-regex';
-import assert from 'assert';
 import { existsSync } from 'fs';
 import npmPackageArg from 'npm-package-arg';
 import path from 'path';
@@ -10,6 +9,7 @@ import split from 'split';
 import { Transform } from 'stream';
 
 import { Logger, PackageManager } from './PackageManager';
+import { PnpmPackageManager } from './PnpmPackageManager';
 import isYarnOfflineAsync from './utils/isYarnOfflineAsync';
 import { findWorkspaceRoot, isUsingYarn, resolvePackageManager } from './utils/nodeWorkspaces';
 
@@ -19,7 +19,7 @@ export type NodePackageManager = 'yarn' | 'npm' | 'pnpm';
  * Disable various postinstall scripts
  * - https://github.com/opencollective/opencollective-postinstall/pull/9
  */
-const disableAdsEnv = { DISABLE_OPENCOLLECTIVE: '1', ADBLOCK: '1' };
+export const DISABLE_ADS_ENV = { DISABLE_OPENCOLLECTIVE: '1', ADBLOCK: '1' };
 
 const ansi = `(?:${ansiRegex().source})*`;
 const npmPeerDependencyWarningPattern = new RegExp(
@@ -63,7 +63,7 @@ export class NpmPackageManager implements PackageManager {
     this.options = {
       env: {
         ...process.env,
-        ...disableAdsEnv,
+        ...DISABLE_ADS_ENV,
       },
       cwd,
       ...(silent
@@ -214,7 +214,7 @@ export class YarnPackageManager implements PackageManager {
     this.options = {
       env: {
         ...process.env,
-        ...disableAdsEnv,
+        ...DISABLE_ADS_ENV,
       },
       cwd,
       ...(silent
@@ -314,101 +314,6 @@ export class YarnPackageManager implements PackageManager {
     const promise = spawnAsync('yarnpkg', args, { ...this.options, ignoreStdio: false });
     if (promise.child.stderr && !this.options.ignoreStdio) {
       promise.child.stderr.pipe(new YarnStderrTransform()).pipe(process.stderr);
-    }
-    return promise;
-  }
-}
-
-export class PnpmPackageManager implements PackageManager {
-  options: SpawnOptions;
-  private log: Logger;
-
-  constructor({ cwd, log, silent }: { cwd: string; log?: Logger; silent?: boolean }) {
-    this.log = log || console.log;
-    this.options = {
-      env: {
-        ...process.env,
-        ...disableAdsEnv,
-      },
-      cwd,
-      ...(silent
-        ? { ignoreStdio: true }
-        : {
-            stdio: ['inherit', 'inherit', 'pipe'],
-          }),
-    };
-  }
-
-  get name() {
-    return 'pnpm';
-  }
-
-  async installAsync() {
-    await this._runAsync(['install']);
-  }
-
-  async addWithParametersAsync(names: string[], parameters: string[]) {
-    if (!names.length) return this.installAsync();
-    await this._runAsync(['add', ...parameters, ...names]);
-  }
-
-  async addAsync(...names: string[]) {
-    await this.addWithParametersAsync(names, []);
-  }
-
-  async addDevAsync(...names: string[]) {
-    if (!names.length) return this.installAsync();
-    await this._runAsync(['add', '--save-dev', ...names]);
-  }
-
-  async addGlobalAsync(...names: string[]) {
-    if (!names.length) return this.installAsync();
-    await this._runAsync(['add', '--global', ...names]);
-  }
-
-  async removeAsync(...names: string[]) {
-    await this._runAsync(['remove', ...names]);
-  }
-
-  async versionAsync() {
-    const { stdout } = await spawnAsync('pnpm', ['--version'], { stdio: 'pipe' });
-    return stdout.trim();
-  }
-
-  async getConfigAsync(key: string) {
-    const { stdout } = await spawnAsync('pnpm', ['config', 'get', key], { stdio: 'pipe' });
-    return stdout.trim();
-  }
-
-  async removeLockfileAsync() {
-    assert(this.options.cwd, 'cwd required for PnpmPackageManager.removeLockfileAsync');
-    const lockfilePath = path.join(this.options.cwd, 'pnpm-lock.yaml');
-    if (existsSync(lockfilePath)) {
-      rimraf.sync(lockfilePath);
-    }
-  }
-
-  async cleanAsync() {
-    assert(this.options.cwd, 'cwd required for PnpmPackageManager.cleanAsync');
-    const nodeModulesPath = path.join(this.options.cwd, 'node_modules');
-    if (existsSync(nodeModulesPath)) {
-      rimraf.sync(nodeModulesPath);
-    }
-  }
-
-  // Private
-  private async _runAsync(args: string[]) {
-    if (!this.options.ignoreStdio) {
-      this.log(`> pnpm ${args.join(' ')}`);
-    }
-
-    // Have spawnAsync consume stdio but we don't actually do anything with it if it's ignored
-    const promise = spawnAsync('pnpm', args, { ...this.options, ignoreStdio: false });
-    if (promise.child.stderr && !this.options.ignoreStdio) {
-      promise.child.stderr
-        .pipe(split(/\r?\n/, (line: string) => line + '\n'))
-        .pipe(new NpmStderrTransform())
-        .pipe(process.stderr);
     }
     return promise;
   }
