+++ packages/package-manager/src/index.ts
@@ -1,5 +1,6 @@
 export * from './PackageManager';
 export * from './NodePackageManagers';
+export * from './PnpmPackageManager';
 export * from './CocoaPodsPackageManager';
 export { default as shouldUseYarn } from './utils/shouldUseYarn';
 export { default as isYarnOfflineAsync } from './utils/isYarnOfflineAsync';
